move the hash functions to the library
