+++ curs_main.c
@@ -37,7 +37,6 @@
 #include "envelope.h"
 #include "format_flags.h"
 #include "globals.h"
-#include "hash.h"
 #include "header.h"
 #include "keymap.h"
 #include "keymap_defs.h"
