+++ group.c
@@ -27,7 +27,6 @@
 #include "group.h"
 #include "address.h"
 #include "globals.h"
-#include "hash.h"
 #include "lib.h"
 #include "lib/lib.h"
 #include "list.h"
