+++ hdrline.c
@@ -37,7 +37,6 @@
 #include "envelope.h"
 #include "format_flags.h"
 #include "globals.h"
-#include "hash.h"
 #include "header.h"
 #include "lib.h"
 #include "lib/lib.h"
