+++ imap/imap.c
@@ -43,7 +43,6 @@
 #include "context.h"
 #include "envelope.h"
 #include "globals.h"
-#include "hash.h"
 #include "header.h"
 #include "imap/imap.h"
 #include "lib.h"
