+++ imap/message.c
@@ -39,7 +39,6 @@
 #include "context.h"
 #include "envelope.h"
 #include "globals.h"
-#include "hash.h"
 #include "header.h"
 #include "imap/imap.h"
 #include "lib.h"
