+++ lib/lib.h
@@ -35,6 +35,7 @@
  * -# @subpage debug
  * -# @subpage exit
  * -# @subpage file
+ * -# @subpage hash
  * -# @subpage md5
  * -# @subpage memory
  * -# @subpage message
@@ -52,6 +53,7 @@
 #include "lib_debug.h"
 #include "lib_exit.h"
 #include "lib_file.h"
+#include "lib_hash.h"
 #include "lib_md5.h"
 #include "lib_memory.h"
 #include "lib_message.h"
