+++ lib/lib_hash.h
@@ -20,8 +20,8 @@
  * this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef _MUTT_HASH_H
-#define _MUTT_HASH_H
+#ifndef _LIB_HASH_H
+#define _LIB_HASH_H
 
 #include <stdbool.h>
 
@@ -91,4 +91,4 @@ struct HashWalkState
 
 struct HashElem *hash_walk(const struct Hash *table, struct HashWalkState *state);
 
-#endif /* _MUTT_HASH_H */
+#endif /* _LIB_HASH_H */
