+++ mh.c
@@ -50,7 +50,6 @@
 #include "copy.h"
 #include "envelope.h"
 #include "globals.h"
-#include "hash.h"
 #include "header.h"
 #include "lib.h"
 #include "lib/lib.h"
