+++ po/POTFILES.in
@@ -27,7 +27,6 @@ from.c
 getdomain.c
 group.c
 handler.c
-hash.c
 hcache/bdb.c
 hcache/gdbm.c
 hcache/hcache.c
@@ -64,6 +63,7 @@ lib/lib_date.c
 lib/lib_debug.c
 lib/lib_exit.c
 lib/lib_file.c
+lib/lib_hash.c
 lib/lib_md5.c
 lib/lib_memory.c
 lib/lib_message.c
