+++ postpone.c
@@ -35,7 +35,6 @@
 #include "envelope.h"
 #include "format_flags.h"
 #include "globals.h"
-#include "hash.h"
 #include "header.h"
 #include "keymap.h"
 #include "keymap_defs.h"
